% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbsas}
\alias{cbsas}
\title{cbsas}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 938 rows and 8 columns.
}
\usage{
cbsas
}
\description{
2019 cbsas at full resolution
}
\keyword{datasets}
