% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{counties}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3233 rows and 18 columns.
}
\usage{
counties
}
\description{
2015 counties at full resolution
}
\keyword{datasets}
