% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties_simplified}
\alias{counties_simplified}
\title{counties_simplified}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3233 rows and 10 columns.
}
\usage{
counties_simplified
}
\description{
2015 counties at 5m resolution
}
\keyword{datasets}
