% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{czs_simplified}
\alias{czs_simplified}
\title{czs_simplified}
\format{
An object of class \code{sf} (inherits from \code{grouped_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 741 rows and 3 columns.
}
\usage{
czs_simplified
}
\description{
1990 CZs at 5m resolution
}
\keyword{datasets}
