% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school.dists}
\alias{school.dists}
\title{school.dists}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 12808 rows and 10 columns.
}
\usage{
school.dists
}
\description{
School districts. Unified districts supplemented with elementary. Year 2019
full resolution.
}
\keyword{datasets}
