% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfr.polys}
\alias{sfr.polys}
\title{sfr.polys}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 20358 rows and 3 columns.
}
\usage{
sfr.polys
}
\description{
Polygons representing SFR zoning status. Built from simplified CTs and
analysis of corelogic data
}
\keyword{datasets}
